Added the capstone project guidelines and requirements document.
This commit adds the project scope to the repository. The PDF contains a detailed description of the project requirements, as well as the specific tasks and objectives that need to be accomplished. It includes information on the system specifications, design requirements, and implementation guidelines. This document will serve as a reference point for the project team throughout the development process.